Update AuthRepository.cs
The Company object was not being tracked by the program, so I decided to attach it in the user creation method in the repository to fix the bug that occurred when an administrator tried to create a user with the same company. Another way to fix this bug would be to change the AsNoTracking() directive to AsTracking() in the company repository. However, due to performance considerations, I chose the other approach.